Add files via upload
The purpose of the project is to create a simple analytics program with User Interaction.

This project is a GUI implementation of the Case Study 15.5 from the Python textbook.  For each section, you will need to have user interaction and the ability to restart at any time.  The user must also be prevented from completing an action that is invalid.  For example, the user cannot explore the data until the data is loaded.
1.	Load the dataset (15.5.1)
2.	Explore the data (15.5.2)
3.	Split the data for training and testing (15.5.4)
4.	Train the data model (15.5.5)
5.	Test the data model (15.5.6)
6.	Visualize the expected vs. predicted (15.5.7)
7.	Create the regression model metrics (15.5.8)